+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 - (Experimental) `C-x <SPC>` (rectangle-mark-mode), #819.
 - Rectangle kill, copy, delete and yank commands (`C-x r k`, `C-x r M-w`, `C-x r d`, `C-x r y`), #825.
 
+### Fix
+- Fix beginning-of-buffer and end-of-buffer not to push mark if mark-mode has already been started, #839.
+
 ## [0.31.0] - 2021-04-20
 ### Fixed
 - Use `isearchExit` with `then` arg instead of `executeCommands` to quit the find widget with key press.